Renames settings.js to settingsmanager.js
