+++ src/themes/SettingsManager.js
